[REFACTOR] Organize metrics
This PR perform one round of reorganization of metrics into
a centralized metrics header.

Also updates the ChatState to include overrides that can be used
in future cases to run chat test.
